requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
hubspot-api-client==12.0.0
idna==3.11
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
urllib3==2.5.0
